Refactored completion tests and restore full async on QuarkusLanguageClient
